rename jest.config.js to index.js in packages
