+++ packages/jest-config/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@webstudio-is/jest-config",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "license": "MIT",
   "dependencies": {
     "@types/jest": "^27.4.1",
