+++ packages/react-sdk/jest.config.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line
-const base = require("@webstudio-is/jest-config/jest.config.js");
+const base = require("@webstudio-is/jest-config");
 
 module.exports = {
   ...base,
